ban on deleting someone else's card

+++ controllers/cards.js
@@ -1,12 +1,14 @@
-const Card = require('../models/card');
+const Card = require("../models/card");
 
 module.exports.getCards = (req, res) => {
   Card.find({})
-    .populate('owner')
+    .populate("owner")
     .then((cards) => {
       res.status(200).send(cards);
     })
-    .catch((err) => res.status(500).send({ message: `Произошла ошибка: ${err}` }));
+    .catch((err) =>
+      res.status(500).send({ message: `Произошла ошибка: ${err}` })
+    );
 };
 
 module.exports.createCard = (req, res) => {
@@ -14,26 +16,38 @@ module.exports.createCard = (req, res) => {
   Card.create({ name, link, owner: req.user._id })
     .then((card) => res.status(200).send({ card }))
     .catch((err) => {
-      if (err.name === 'ValidationError') {
+      if (err.name === "ValidationError") {
         return res.status(400).send({ message: `Ошибка валидации: ${err}` });
       }
       return res.status(500).send({ message: `Произошла ошибка: ${err}` });
     });
 };
 
-module.exports.deleteCard = (req, res) => {
-  Card.findByIdAndRemove(req.params.cardId)
+module.exports.deleteCard = (req, res, next) => {
+  Card.findById(req.params.cardId)
     .then((card) => {
       if (!card) {
-        return res.status(404).send({ message: 'Карточка не найдена' });
-      }
-      return res.status(200).send({ data: card });
-    })
-    .catch((err) => {
-      if (err.name === 'CastError') {
-        res.status(400).send({ message: `Некорректно указаны данные карточки: ${err}` });
+        res.status(404).send({ message: 'Card not found.' });
+        // throw new NotFoundError('Card not found')
+      } else if (card.owner == req.user._id) {
+        Card.findByIdAndRemove(req.params.cardId)
+          .then((card) => {
+            if (!card) {
+              res.status(404).send({ message: 'Card not found!' });
+            } else {
+              res.send({ data: card });
+            }
+          })
+          .catch((error) => {
+            if (error.name === 'CastError') {
+              res.status(404).send({ message: 'Card not found!!' });
+            } else {
+              res.status(500).send({ message: 'Произошла ошибка' });
+            }
+          });
       } else {
-        res.status(500).send({ message: `Произошла ошибка: ${err}` });
+        res.status(404).send({ message: 'You cant delete this card.' });
       }
-    });
+    })
+    .catch(next)
 };
